Dynamically generate board
Now that jQuery is present in the project, we can start by generating
the HTML and CSS for the board entirely in JavaScript. This lets us
easily change the size of the board down the road if we want to.

However, if we were to view this in a browser we would see nothing. This
is because by the time our JavaScript is loaded, the div that represents
the board hasn't been rendered yet, so when our JavaScript tries to look
it up, it will find nothing. Frustratingly, we won't see any errors in
the JavaScript console since jQuery silently fails if it can't find
anything, but that is what is happening here.