Reveal mines when clicked
Here we add a click handler to each cell. If the cell is a mine, then we
reveal all of the cells (whether a mine or not) and end the game.
Otherwise, we reveal just that cell. The cell gets a number indicating
how many cells that are adjacent to that cell are mines.

To visually identify whether a cell is a mine, we apply a CSS class. To
be able to look up whether a cell is a mine, we add a key to the `data`.

We also have to be able to access the X and Y coordinate of a certain
cell, so we store it in `data`.

Finally, we have to be able to look up the neighboring cells of a
particular cell. This means we need to be able to look up cells by their
X and Y coordinate, so we use CSS classes for this. But, notice how
cumbersome it is to do this! We're essentially using the DOM as a lookup
table, which it was never intended to do. Surely there's a better way?