Refactor!
At this point, our code is a little messy. For starters, notice how we
have comments all over the place. They're there for a reason: to help us
remember some of the more complicated pieces. But think about it --
if we have pieces that are so complicated they need to be pointed out,
perhaps they need to be separated from the other parts of the code.
So a lot of this commit is separating pieces.

In addition, we can clean up some questionable practices. If we run our
code through JSHint, which is a popular "linting" tool, it will give us
a few warnings. Two of them are about variables being redefined.
However, one of them is due to the the click handler (i.e., what happens
when you click on a cell) being in a loop. Although this doesn't happen
to cause any issues in our project, this *could* cause issues, and is
therefore a poor design, because there are variables defined during or
outside of the loop that are also being accessed inside of the function
which is the click handler, and therefore the values of those variables
could change unexpectedly by the time this function is called.

So, we can do some things to clean up our code. Take a look at what
we've done:

* We've extracted the code that generates the board to its own method.
* We've extracted the code that generates a single cell to its own
  method. This removes a warning from JSHint: "Don't make functions
  within a loop."
* We've extracted the code that determines whether a mine is at a
  particular location.
* We've extracted the click handler mentioned above.
* We've extracted code that ends the game (by identifying all cells) to
  its own method. As a result, this removes the warning from JSHint:
  "'cell' is already defined".
* We've extracted code that finds adjacent mines and counts the number
  of mines in a given area. These two operations remove the warning from
  JSHint: "'k' is already defined".
* Finally, we've wrapped everything in an IIFE (Immediately Invoked
  Function Expression), as well as code that configures the game itself
  such as setting the board size and number of mines. This ensures that
  no variables leak out into the global scope. In addition, all
  functions now take arguments for variables that they need to use
  instead of assuming those variables are accessible in the parent
  scope. This is a better design because we are better aware which
  state each variable is in.