Refactor functions into Board class
Nothing much happening in this commit -- we're mainly taking all of the
functions we had before and putting them inside of a Board class.
Additionally, we take the configuration variables that we set at the end
of the file and pass them to the Board constructor as options.

Additionally, we also rename `generateBoard` to `generateCells`, and
group `generateMineLocations` and `generateCells` under a `generate`
function.

Notice that most of the functions here are defined directly in the Board
constructor, and only `generate` is an actual method you can call on a
Board object. We don't want those other functions to be publicly
callable, because otherwise, Board would be exposing a bunch of methods
to users of this class who don't need to call them directly. Instead, we
want to keep our API at a reasonable size.

If we were writing this code in Ruby, we would achieve this by create
methods under a `private` section. But JavaScript doesn't have a
built-in way to do this, since in JavaScript, methods are merely
properties of an object that are functions, and there is no such thing
as private properties.

So if we wanted to designate methods as private, what do we do? Here are
a few ways people have done this:

1. By defining them as functions in the constructor, as we've done. The
   scope of these functions are limited to Board, since it is itself a
   function. We can call these functions anywhere inside of Board, but
   we can't outside.

2. By defining them as methods on the class, but name them in such a
   way, such as starting the name with an underscore, that users will
   notice and hopefully treat them differently.

3. By defining them as methods on the class, but using Symbol to give
   their names something unique and non-guessable. This technique isn't
   used a lot and is slightly advanced, but you can read [here] for more
   information, if you're interested.

[here]: https://curiosity-driven.org/private-properties-in-javascript