Extract Cell class
Here we're continuing the task of converting functions into objects.
When we added the ability to find neighboring cells, we mentioned that
we keep and look up the location of those cells using CSS classes.
This is, admittedly, a poor use of CSS classes: we're using them to
store data, when they were really only invented to change presentation.
It turns out we have a much better place to store data: code, and more
specifically, properties on objects.

So here, we're cleaning some things up by moving some data to a Cell
class -- not only the coordinates of a cell, but the element the cell is
backed by (along with its click handler) and whether that cell is a
mine. Also, we add a couple of methods to reveal the cell and access the
cells adjacent to the cell.