Switch everything to a model/view structure
In this commit, we are splitting off the core logic of the game (which
cells have been revealed, how many mines are around dangerous cells)
from the presentation logic (what the board looks like on screen
depending on the state of the game). In other words, we're switching
to a model/view structure.

The way that the game works now is this:

* There are two models: board and cell. A board holds all of the cells
  (grouped by row).
* There are three views: board view, row view, and cell view. Each
  view is associated with an element: cell views represent an
  individual cell div, row views represent row divs which wrap cell
  divs, and the board view represents the entire board div that wraps
  the row divs which wrap the cell divs.
* Views have access to models, but not the other way around.
* Clicking on a cell changes the state of that cell, and possibly
  the cells around it, from not-revealed to revealed. The entire view
  is then re-rendered based off the model.
* Here, the view is actually acting as a controller as well, since
  the click event is handled by the view. That doesn't usually happen
  -- the controller is separate.
* Re-rendering the entire board isn't as bad as it sounds. We first
  empty the board div entirely, append empty row divs, ask each row
  div to re-render and thus populate itself, which will then append
  cell divs to the row and ask each cell div to re-render and populate
  itself.

This commit was difficult, and took a long time to write (about 5x as
long as the other commits). There were three things that made it
difficult:

* Figuring out which levels of abstractions to create
* Figuring out the interface in each class -- which methods to
  create, and which objects were supposed to access which other objects
* Making everything easy to test

Because of this, this commit went through a number of iterations.

At first, there was only two levels, BoardView and CellView, where
the CellViews were kept inside of an array inside of BoardView. I
decided to keep that array private (because only a board view
should be able to access its cell views), but I quickly realized that
with an array, I couldn't verify that the board view called various
methods (#activate, #appendTo, and #render) on the cell views.

So then I created a CellViewCollection, put it outside of BoardView,
and brought in as a constructor argument. But that merely shifted
the problem around, because while I could make assertions on the
CellViewCollection from BoardView, how was I supposed to test
CellViewCollection itself? I tried placing a #toArray method on the
collection -- that let me get at the cells inside. But that didn't
make sense because I never used #toArray anywhere but tests. It took
me a while, but then realized that if I simply had a #forEach method
(which *was* used outside the tests) I could then access what was
inside.

I also realized, as I was testing out my changes visually, that I had
to have a middle layer (row), because cell divs are actually wrapped
in row divs. That meant I had to have a RowView, and
CellViewCollection could change to simply ViewCollection. This also
meant I had to have a way to make a ViewCollection that consisted
of rows and a ViewCollection that consisted of cells. This took a
couple of tries to get right but I finally landed on a design that
made it easy to test. At the same time, ViewCollection was moved back
into BoardView.

Having three layers made things a lot more complex, and I had to keep
track of what things talked to what other things. Do cells have
direct access to the board? (Yes, because there isn't a Row model
in between.) Do cell views have direct access to the board view?
(No, because there *is* a RowView in between.)

Was the complexity worth it? Probably not in this case. Minesweeper
is a simple game and doesn't take a lot of code to write. But for
larger games or even sites, a model/view structure makes a lot of
sense. There will probably be a lot more logic, and keeping that
separate from what appears on screen makes things easier to manage
-- provided that all of the pieces fit together in a reasonable way.